+++ src/package/themes/default/components/checkbox.ts
@@ -4,7 +4,7 @@ export default {
     borderRadius: 'var(--vf-radii-sm)',
     checkBorderRadius: '2px',
     borderColor: 'var(--vf-border-color)',
-    backgroundColor: 'var(--vf-bg-color)',
+    backgroundColor: 'var(--vf-controls-background-color)',
     checkedBackgroundColor: 'var(--vf-blue)',
     checkedBorderColor: 'var(--vf-blue)',
     checkColor: '#ffffff',